약관 페이지 반응형 디자인 적용
