+++ homepage/app/_components/index.jsx
@@ -1,5 +1,4 @@
 // components export
 
-export { default as Header } from "./Header"
-export { default as Policy } from "./Policy"
-export { default as Section } from "./Policy/Section"
+export { default as Header } from "./Header";
+export { default as Policy } from "./Policy";
