+++ homepage/app/data/policiesData.js
@@ -1,80 +1,78 @@
-import Link from "next/link"
-
 const policiesData = [
-    {
-        title: "1. 링크드아웃 개인정보 처리방침",
-        content: [
-            `<p class="mb-2"> 
+  {
+    title: "1. 링크드아웃 개인정보 처리방침",
+    content: [
+      `<p class="mb-2"> 
          &ldquo;개인정보 처리방침&rdquo;이란 이용자가 안심하고 서비스를 이용할 수 있도록 회사가 준수해야 할 
          지침을 의미하며, 링크드아웃은 개인정보처리자가 준수하여야 하는 대한민국의 관계 법령 및 개인정보보호 
          규정, 가이드라인을 준수하여 개인정보 처리방침을 제공합니다. 
       </p>`,
-            `<p class="mb-4"> 
+      `<p class="mb-4"> 
          링크드아웃은 이용자의 동의를 기반으로 개인정보를 수집&middot;이용 및 제공하고 있습니다. 이용자의 
          권리(개인정보 자기결정권)를 적극적으로 보장하기 위해 개인정보 처리방침을 알기 쉽게 제공할 수 있도록 
          다양한 노력을 기울이고 있으며, 이러한 노력의 일환으로 링크드아웃의 주요 개인정보 처리 관련 정보를 
          라벨링으로 제공합니다. 
       </p>`,
-            {
-                type: "images",
-                images: [
-                    {
-                        src: "/images/개인정보.png",
-                        width: 100,
-                        height: 100,
-                        alt: "개인정보",
-                        caption: "개인정보",
-                    },
-                    {
-                        src: "/images/처리목적.png",
-                        width: 100,
-                        height: 100,
-                        alt: "처리목적",
-                        caption: "처리목적",
-                    },
-                    {
-                        src: "/images/제3자제공.png",
-                        width: 100,
-                        height: 100,
-                        alt: "제3자 제공",
-                        caption: "제3자 제공",
-                    },
-                    {
-                        src: "/images/처리위탁.png",
-                        width: 120,
-                        height: 120,
-                        alt: "처리위탁",
-                        caption: "처리위탁",
-                    },
-                    {
-                        src: "/images/정보주체의권리의무.png",
-                        width: 100,
-                        height: 100,
-                        alt: "정보주체의 권리의무",
-                        caption: "정보주체의 권리의무",
-                    },
-                    {
-                        src: "/images/고충처리부서.png",
-                        width: 100,
-                        height: 100,
-                        alt: "고충처리 부서",
-                        caption: "고충처리 부서",
-                    },
-                ],
-            },
+      {
+        type: "images",
+        images: [
+          {
+            src: "/images/개인정보.png",
+            width: 100,
+            height: 100,
+            alt: "개인정보",
+            caption: "개인정보",
+          },
+          {
+            src: "/images/처리목적.png",
+            width: 100,
+            height: 100,
+            alt: "처리목적",
+            caption: "처리목적",
+          },
+          {
+            src: "/images/제3자제공.png",
+            width: 100,
+            height: 100,
+            alt: "제3자 제공",
+            caption: "제3자 제공",
+          },
+          {
+            src: "/images/처리위탁.png",
+            width: 120,
+            height: 120,
+            alt: "처리위탁",
+            caption: "처리위탁",
+          },
+          {
+            src: "/images/정보주체의권리의무.png",
+            width: 100,
+            height: 100,
+            alt: "정보주체의 권리의무",
+            caption: "정보주체의 권리의무",
+          },
+          {
+            src: "/images/고충처리부서.png",
+            width: 100,
+            height: 100,
+            alt: "고충처리 부서",
+            caption: "고충처리 부서",
+          },
         ],
-    },
-    {
-        title: "2. 개인정보 수집",
-        content: [
-            `<h3 class="mb-2 font-bold text-l">서비스 제공을 위한 필요 최소한의 개인정보를 수집합니다.</h3>`,
-            `<p class="mb-4"> 
+      },
+    ],
+  },
+  {
+    title: "2. 개인정보 수집",
+    content: [
+      `<h3 class="mb-2 font-bold text-l">서비스 제공을 위한 필요 최소한의 개인정보를 수집합니다.</h3>`,
+      `<p class="mb-4"> 
          회원 가입 시 또는 서비스 이용 과정에서 홈페이지 또는 개별 어플리케이션이나 프로그램 등을 통해 서비스 
          제공을 위해 필요 최소한의 개인정보를 수집하고 있습니다. 서비스 제공을 위해 반드시 필요한 최소한의 
          정보를 필수항목으로, 그 외 특화된 서비스를 제공하기 위해 추가 수집하는 정보는 선택항목으로 동의를 
          받고 있으며, 선택항목에 동의하지 않은 경우에도 서비스 이용 제한은 없습니다. 
       </p>`,
-            `<div class="p-4">
+      `<div class="p-4">
           <p class="mb-2 font-semibold">[링크드아웃 계정 가입 시]</p>
           <p class="mb-2">
             [필수] 이메일, 비밀번호, 이름(닉네임), 프로필사진, 친구목록, 서비스 이용내역, 서비스 내 구매 및 
@@ -82,55 +80,55 @@ const policiesData = [
           </p>
           <p>[선택] 생년월일, 성별, 배송지정보(수령인명, 배송지주소, 전화번호)</p>
        </div>`,
-            `<div class="p-4">
+      `<div class="p-4">
           <p class="mb-4 font-semibold">[고객상담 시]</p>
           <p>고객센터로 문의 및 상담 시 상담 처리를 위한 추가적인 정보를 수집할 수 있습니다.</p>
        </div>`,
-            `<h3 class="mb-2 font-bold text-l">개인정보를 수집하는 방법은 아래와 같습니다.</h3>`,
-            `<p class="mb-4"> 
+      `<h3 class="mb-2 font-bold text-l">개인정보를 수집하는 방법은 아래와 같습니다.</h3>`,
+      `<p class="mb-4"> 
          개인정보를 수집하는 경우에는 원칙적으로 사전에 이용자에게 해당 사실을 알리고 동의를 구하고 있으며, 아래와 같은 방법을 통해 개인
 정보를 수집합니다.
       </p>`,
-            `<ul class="pl-4 mb-10 list-disc sm:pl-10">
+      `<ul class="pl-4 mb-10 list-disc sm:pl-10">
           <li class="my-4">회원가입 및 서비스 이용 과정에서 이용자가 개인정보 수집에 대해 동의를 하고 직접 정보를 입력하는 경우</li>
           <li class="my-4">
             고객센터를 통한 상담 과정에서 웹페이지, 메일, 팩스, 전화 등
           </li>
        </ul>`,
-            `<h3 class="mb-2 font-bold text-l">서비스 이용 과정에서 단말기정보, IP주소, 쿠키, 서비스 이용 내역 등의 정보가 자동으로 생성되어 수집될 수 있습니다.</h3>`,
-            `<p class="mb-4"> 
+      `<h3 class="mb-2 font-bold text-l">서비스 이용 과정에서 단말기정보, IP주소, 쿠키, 서비스 이용 내역 등의 정보가 자동으로 생성되어 수집될 수 있습니다.</h3>`,
+      `<p class="mb-4"> 
          서비스 이용 내역이란 서비스 이용 과정에서 자동화된 방법으로 생성되거나 이용자가 입력한 정보가 송수신되면서 링크드아웃 서버에
 자동으로 기록 및 수집될 수 있는 정보를 의미합니다. 이와 같은 정보는 다른 정보와의 결합 여부, 처리하는 방식 등에 따라 개인정보에
 해당할 수 있고 개인정보에 해당하지 않을 수도 있습니다.
       </p>`,
-            `<p class="mb-4"> 
+      `<p class="mb-4"> 
          서비스 이용 내역에는 이용자가 입력 및 공유한 콘텐츠, 이용자가 입력한 검색어, 방문 및 접속기록, 서비스 부정이용 기록, 위치정보 등
 이 포함될 수 있습니다. 링크드아웃은 서비스 이용기록 등의 정보를 서비스 제공 목적으로 처리할 수 있으며 필요한 경우에는 이용자의
 추가 동의 등을 받고 이용할 수 있습니다.
       </p>`,
-        ],
-        figures: [
-            {
-                imageSrc: "/images/개인정보.png",
-                alt: "개인정보",
-                caption: "개인정보",
-                description:
-                    "링크드아웃 계정 가입 시 서비스 제공을 위해 필요한 최소한의 개인정보를 수집하고 있습니다. 서비스 이용 시 특화된 기능 이용을 위해 이용자의 동의를 받고 추가적인 개인정보를 수집할 수 있습니다.",
-            },
-            {
-                imageSrc: "/images/자동화수집.png",
-                alt: "자동화 수집",
-                caption: "자동화 수집",
-                description:
-                    "링크드아웃 서비스 이용 과정에서 단말기정보, IP주소, 쿠키, 방문일시, 부정이용기록, 서비스 이용 기록 등의 정보가 자동으로 생성되어 수집될 수 있습니다.",
-            },
-        ],
-    },
-    {
-        title: "3. 개인정보 이용",
-        content: [
-            `<p class="mb-4 font-black">회원관리, 서비스 제공&middot;개선, 신규 서비스 개발 등을 위해 이용합니다.</p>`,
-            `<ul class="pl-4 mb-4 list-disc sm:pl-10">
+    ],
+    figures: [
+      {
+        imageSrc: "/images/개인정보.png",
+        alt: "개인정보",
+        caption: "개인정보",
+        description:
+          "링크드아웃 계정 가입 시 서비스 제공을 위해 필요한 최소한의 개인정보를 수집하고 있습니다. 서비스 이용 시 특화된 기능 이용을 위해 이용자의 동의를 받고 추가적인 개인정보를 수집할 수 있습니다.",
+      },
+      {
+        imageSrc: "/images/자동화수집.png",
+        alt: "자동화 수집",
+        caption: "자동화 수집",
+        description:
+          "링크드아웃 서비스 이용 과정에서 단말기정보, IP주소, 쿠키, 방문일시, 부정이용기록, 서비스 이용 기록 등의 정보가 자동으로 생성되어 수집될 수 있습니다.",
+      },
+    ],
+  },
+  {
+    title: "3. 개인정보 이용",
+    content: [
+      `<p class="mb-4 font-black">회원관리, 서비스 제공&middot;개선, 신규 서비스 개발 등을 위해 이용합니다.</p>`,
+      `<ul class="pl-4 mb-4 list-disc sm:pl-10">
           <li class="my-4">회원 식별/가입의사 확인, 본인/연령 확인</li>
           <li class="my-4">
             14세 미만 아동의 개인정보 수집 시 법정 대리인 동의여부 확인, 법정 대리인 권리행사 시 본인 확인
@@ -152,105 +150,105 @@ const policiesData = [
           </li>
           <li class="my-4">서비스 이용 기록, 접속 빈도 및 서비스 이용에 대한 통계</li>
        </ul>`,
-            `<p class="mt-12 mb-8 font-black">개인정보의 추가적인 이용・제공을 하는 경우가 있습니다.</p>`,
-            `<p class="mb-6">
+      `<p class="mt-12 mb-8 font-black">개인정보의 추가적인 이용・제공을 하는 경우가 있습니다.</p>`,
+      `<p class="mb-6">
          수집 목적과 합리적으로 관련된 범위에서는 법령에 따라 이용자의 동의 없이 개인정보를 이용하거나 
          제3자에게 제공할 수 있습니다. 이 때 &apos;당초 수집 목적과 관련성이 있는지, 수집한 정황 또는 처리 
          관행에 비추어 볼 때 개인정보의 추가적인 이용 또는 제공에 대한 예측 가능성이 있는지, 이용자의 이익을 
          부당하게 침해하는지, 가명처리 또는 암호화 등 안전성 확보에 필요한 조치를 하였는지&apos;를 종합적으로 
          고려합니다.
       </p>`,
-            `<p class="mb-12">
+      `<p class="mb-12">
          링크드아웃은 수집한 개인정보를 특정 개인을 알아볼 수 없도록 가명처리하여 통계작성, 과학적 연구, 
          공익적 기록보존 등을 위하여 처리할 수 있습니다. 이 때 가명정보는 재식별되지 않도록 추가정보와 
          분리하여 별도 저장・관리하고 필요한 기술적・관리적 보호조치를 취합니다.
       </p>`,
-        ],
-        figures: [
-            {
-                imageSrc: ["/images/처리항목.png", "/images/처리목적.png"],
-                alt: "처리항목 및 처리목적",
-                caption: "처리항목 및 처리목적",
-                description:
-                    "이용자의 동의를 받아 수집한 개인정보 항목은 회원 관리 등 서비스 제공을 위해 필요한 목적으로 처리합니다. 링크드아웃에서 처리하는 개인정보 항목 및 이용 목적은 개인정보 처리방침을 통해 상시 공개하고 있습니다.",
-            },
-            {
-                imageSrc: "/images/추가적이용.png",
-                alt: "추가적 이용",
-                caption: "추가적 이용",
-                description:
-                    "당초 수집 목적과 관련성, 수집한 정황 또는 처리 관행 등을 종합적으로 고려하여 수집 목적과 합리적으로 관련된 범위에서 추가적으로 개인정보를 이용 및 제공할 수 있습니다.",
-            },
-            {
-                imageSrc: "/images/가명정보처리.png",
-                alt: "가명정보 처리",
-                caption: "가명정보 처리",
-                description:
-                    "수집한 개인정보를 특정 개인을 알아볼 수 없도록 가명처리하여 통계작성, 과학적 연구, 공익적 기록보존 등을 위해 처리할 수 있습니다.",
-            },
-        ],
-    },
-    {
-        title: "4. 개인정보 제공",
-        content: [
-            `<h3 class="mb-4 font-bold">
+    ],
+    figures: [
+      {
+        imageSrc: ["/images/처리항목.png", "/images/처리목적.png"],
+        alt: "처리항목 및 처리목적",
+        caption: "처리항목 및 처리목적",
+        description:
+          "이용자의 동의를 받아 수집한 개인정보 항목은 회원 관리 등 서비스 제공을 위해 필요한 목적으로 처리합니다. 링크드아웃에서 처리하는 개인정보 항목 및 이용 목적은 개인정보 처리방침을 통해 상시 공개하고 있습니다.",
+      },
+      {
+        imageSrc: "/images/추가적이용.png",
+        alt: "추가적 이용",
+        caption: "추가적 이용",
+        description:
+          "당초 수집 목적과 관련성, 수집한 정황 또는 처리 관행 등을 종합적으로 고려하여 수집 목적과 합리적으로 관련된 범위에서 추가적으로 개인정보를 이용 및 제공할 수 있습니다.",
+      },
+      {
+        imageSrc: "/images/가명정보처리.png",
+        alt: "가명정보 처리",
+        caption: "가명정보 처리",
+        description:
+          "수집한 개인정보를 특정 개인을 알아볼 수 없도록 가명처리하여 통계작성, 과학적 연구, 공익적 기록보존 등을 위해 처리할 수 있습니다.",
+      },
+    ],
+  },
+  {
+    title: "4. 개인정보 제공",
+    content: [
+      `<h3 class="mb-4 font-bold">
          링크드아웃은 이용자의 별도 동의가 있거나 법령에 규정된 경우를 제외하고는 이용자의 개인정보를 
          제3자에게 제공하지 않습니다.
       </h3>`,
-            `<p class="mb-4">
+      `<p class="mb-4">
          링크드아웃은 이용자의 사전 동의 없이 개인정보를 제3자에게 제공하지 않습니다. 다만, 이용자가 
          링크드아웃계정 로그인 서비스를 이용하거나 외부 제휴사 등의 서비스를 이용하는 경우 필요한 범위 내에서 
          이용자의 동의를 얻은 후에 개인정보를 제3자에게 제공하고 있습니다.
       </p>`,
-            `<p class="mb-4">
+      `<p class="mb-4">
          또한 해외 서비스 이용을 위해 이용자의 동의를 받아 국외로 개인정보가 제공될 수 있으며, 국외 이전 시 
          제공받는 자 및 이전국가 등에 대해 고지하여 별도 동의를 받고 있습니다.
       </p>`,
-            `<p class="mb-4">
+      `<p class="mb-4">
          링크드아웃은 재난, 감염병, 급박한 생명・신체 위험을 초래하는 사건사고, 급박한 재산 손실 등의 
          긴급상황이 발생하는 경우 정보주체의 동의 없이 관계기관에 개인정보를 제공할 수 있습니다.
       </p>`,
-            `<h3 class="mt-8 font-bold">
+      `<h3 class="mt-8 font-bold">
          서비스 제공을 위해 아래와 같은 업무를 위탁합니다.
       </h3>`,
-            `<p class="mb-12">
+      `<p class="mb-12">
         서비스 제공을 위해 필요한 경우 개인정보 처리 업무 중 일부를 외부에 위탁할 수 있습니다. 위탁받은 업체가 위탁받은 업무 목적 외로 개인정보를 처리하는 것을 제한하고, 기술적・관리적 보호조치 적용 및 재위탁 제한 등 위탁받은 업체의 개인정보 보호 관련 법령 준수 여부를 관리·감독하고 있습니다. 이용자는 고객센터를 통해 개인정보의 국외 이전을 거부할 수 있습니다. 이용자가 개인정보의 국외 이전을 거부하는 경우, 개인정보 국외이전이 필수적으로 수반되는 서비스의 이용이 제한될 수 있습니다.</p>`,
-        ],
-        figures: [
-            {
-                imageSrc: "/images/제3자제공.png",
-                alt: "제3자 제공",
-                caption: "제3자 제공",
-                description:
-                    "외부 제휴사 등의 서비스 이용 시 이용자에게 제3자 제공 동의를 받은 후 개인정보를 제공하고 있습니다. 개인정보가 제공될 수 있는 제3자 업체 목록은 링크드아웃 개인 정보 처리방침을 통해 확인하실 수 있습니다.",
-            },
-            {
-                imageSrc: "/images/처리위탁.png",
-                alt: "처리위탁",
-                caption: "처리위탁",
-                description:
-                    "링크드아웃 서비스 제공에 있어 반드시 필요한 업무 중 일부를 외부 업체에서 수행할 수 있도록 개인정보를 위탁하고 있습니다. 위탁 시 위탁받은 업체가 개인정보보호 관련 법령을 준수할 수 있도록 정기적으로 관리・감독하고 있습니다.",
-            },
-            {
-                imageSrc: "/images/국외이전.png",
-                alt: "국외이전",
-                caption: "국외이전",
-                description:
-                    "국외 제휴 업체의 서비스를 이용 시 사전동의를 받아 이용자의 개인정보가 국외로 제공될 수 있습니다. 또한 해외 SMS 발송업무 등을 목적으로 국외 업체에 특정 개인정보 처리를 위탁하고 있습니다",
-            },
-        ],
-    },
-    {
-        title: "5. 개인정보 파기",
-        content: [
-            `<p class="mb-4 font-bold">수집 및 이용목적이 달성된 경우 수집한 개인정보는 지체없이 파기합니다.</p>`,
-            `<p class="mb-4">
+    ],
+    figures: [
+      {
+        imageSrc: "/images/제3자제공.png",
+        alt: "제3자 제공",
+        caption: "제3자 제공",
+        description:
+          "외부 제휴사 등의 서비스 이용 시 이용자에게 제3자 제공 동의를 받은 후 개인정보를 제공하고 있습니다. 개인정보가 제공될 수 있는 제3자 업체 목록은 링크드아웃 개인 정보 처리방침을 통해 확인하실 수 있습니다.",
+      },
+      {
+        imageSrc: "/images/처리위탁.png",
+        alt: "처리위탁",
+        caption: "처리위탁",
+        description:
+          "링크드아웃 서비스 제공에 있어 반드시 필요한 업무 중 일부를 외부 업체에서 수행할 수 있도록 개인정보를 위탁하고 있습니다. 위탁 시 위탁받은 업체가 개인정보보호 관련 법령을 준수할 수 있도록 정기적으로 관리・감독하고 있습니다.",
+      },
+      {
+        imageSrc: "/images/국외이전.png",
+        alt: "국외이전",
+        caption: "국외이전",
+        description:
+          "국외 제휴 업체의 서비스를 이용 시 사전동의를 받아 이용자의 개인정보가 국외로 제공될 수 있습니다. 또한 해외 SMS 발송업무 등을 목적으로 국외 업체에 특정 개인정보 처리를 위탁하고 있습니다.",
+      },
+    ],
+  },
+  {
+    title: "5. 개인정보 파기",
+    content: [
+      `<p class="mb-4 font-bold">수집 및 이용목적이 달성된 경우 수집한 개인정보는 지체없이 파기합니다.</p>`,
+      `<p class="mb-4">
          수집 및 이용 목적의 달성 또는 회원 탈퇴 등 파기 사유가 발생한 경우 개인정보의 형태를 고려하여 
          파기방법을 정합니다. 전자적 파일 형태인 경우 복구 및 재생되지 않도록 안전하게 삭제하고, 그 밖에 
          기록물, 인쇄물, 서면 등의 경우 분쇄하거나 소각하여 파기합니다.
       </p>`,
-            `<p class="mb-4">내부 방침에 따라 일정 기간 보관 후 파기하는 정보는 아래와 같습니다.</p>`,
-            `<ul class="pl-4 mb-4 sm:pl-6">
+      `<p class="mb-4">내부 방침에 따라 일정 기간 보관 후 파기하는 정보는 아래와 같습니다.</p>`,
+      `<ul class="pl-4 mb-4 sm:pl-6">
           <li>
             1. 아래 정보는 탈퇴일로부터 최대 1년간 보관 후 파기합니다.
             <ul class="ml-6 list-disc sm:pl-6">
@@ -263,31 +261,31 @@ const policiesData = [
           </li>
           <li class="my-4">2. 권리침해 신고 및 유해정보 신고 이력은 5년간 보관 후 파기합니다.</li>
        </ul>`,
-            `<p class="mb-4">이 외에 법령에 따라 개인정보를 처리합니다.</p>`,
-        ],
-        figures: [
-            {
-                imageSrc: ["/images/보유기간.png", "/images/파기.png"],
-                alt: "보유기간 및 파기",
-                caption: "보유기간 및 파기",
-                description:
-                    "수집 및 이용 목적의 달성 또는 회원 탈퇴 등 파기사유가 발생한 개인정보는 안전하게 파기합니다. 수집된 개인정보는 사전에 안내된 보유기간 동안만 보관 후 파기하며, 관계법령에 따라 일정보관이 필요한 정보는 별도로 보관합니다.",
-            },
-        ],
-    },
-    {
-        title: "6. 개인위치정보의 처리",
-        content: [
-            `<p class="mb-4 font-bold">
+      `<p class="mb-4">이 외에 법령에 따라 개인정보를 처리합니다.</p>`,
+    ],
+    figures: [
+      {
+        imageSrc: ["/images/보유기간.png", "/images/파기.png"],
+        alt: "보유기간 및 파기",
+        caption: "보유기간 및 파기",
+        description:
+          "수집 및 이용 목적의 달성 또는 회원 탈퇴 등 파기사유가 발생한 개인정보는 안전하게 파기합니다. 수집된 개인정보는 사전에 안내된 보유기간 동안만 보관 후 파기하며, 관계법령에 따라 일정보관이 필요한 정보는 별도로 보관합니다.",
+      },
+    ],
+  },
+  {
+    title: "6. 개인위치정보의 처리",
+    content: [
+      `<p class="mb-4 font-bold">
          링크드아웃은 위치정보의 보호 및 이용 등에 관한 법률에 따라 아래와 같이 개인위치정보를 처리합니다.
       </p>`,
-            `<p class="mb-4 font-bold">
+      `<p class="mb-4 font-bold">
          <a href="/위치기반서비스_이용약관.pdf" target="_blank" rel="noopener noreferrer" class="text-blue-800">
            링크드아웃 위치정보 이용약관 제4조
          </a>
          에 따른 서비스의 제공을 위해 개인위치정보를 보유할 수 있습니다.
       </p>`,
-            `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
+      `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
           <li class="mb-4">
             위치기반서비스 이용 및 제공 목적 달성한 때에는 지체없이 개인위치정보를 파기합니다.
           </li>
@@ -300,8 +298,8 @@ const policiesData = [
             개인위치정보를 보유할 수 있습니다.
           </li>
        </ul>`,
-            `<p class="mb-4 font-bold">개인위치정보의 수집 및 이용목적이 달성되면 지체없이 파기합니다.</p>`,
-            `<div>
+      `<p class="mb-4 font-bold">개인위치정보의 수집 및 이용목적이 달성되면 지체없이 파기합니다.</p>`,
+      `<div>
           <p class="mb-4">
             수집 및 이용 목적의 달성 또는 회원 탈퇴 등 개인위치정보 처리목적이 달성된 경우, 개인위치정보를 
             복구 및 재생되지 않도록 안전하게 삭제합니다. 다만, 다른 법령에 따라 보관해야 하는 등 정당한 
@@ -312,8 +310,8 @@ const policiesData = [
             6개월간 보관합니다.
           </p>
        </div>`,
-            `<p class="my-8 font-bold">이용자의 사전 동의 없이 개인위치정보를 제3자에게 제공하지 않습니다.</p>`,
-            `<div>
+      `<p class="my-8 font-bold">이용자의 사전 동의 없이 개인위치정보를 제3자에게 제공하지 않습니다.</p>`,
+      `<div>
           <p class="mb-4">
             링크드아웃은 이용자의 동의 없이 개인위치정보를 제3자에게 제공하지 않으며, 제3자에게 제공하는 
             경우에는 제공받는 자 및 제공목적을 사전에 이용자에게 고지하고 동의를 받습니다.
@@ -323,10 +321,10 @@ const policiesData = [
             자, 제공일시 및 제공목적을 즉시 통지합니다.
           </p>
        </div>`,
-            `<p class="my-8 font-bold">
+      `<p class="my-8 font-bold">
          8세 이하의 아동 등의 보호를 위해 필요한 경우 보호의무자의 권리는 아래와 같습니다.
       </p>`,
-            `<div>
+      `<div>
           <p class="mb-4">
             위치정보법 제26조제1항에 따라 &lsquo;8세 이하의 아동&rsquo;, &lsquo;피성년후견인&rsquo;,
             &lsquo;장애인복지법에 따른 장애인에 해당하는 이용자&rsquo;의 생명 또는 신체 보호를 위하여
@@ -345,56 +343,56 @@ const policiesData = [
             을 참고하시기 바랍니다.
           </p>
        </div>`,
-        ],
-        figures: [
-            {
-                imageSrc: ["/images/보유기간.png", "/images/파기.png"],
-                alt: "보유기간 및 파기",
-                caption: "보유기간 및 파기",
-                description:
-                    "수집 및 이용 목적의 달성 또는 회원 탈퇴 등 파기사유가 발생한 개인정보는 안전하게 파기합니다. 수집된 개인정보는 사전에 안내된 보유기간 동안만 보관 후 파기하며, 관계법령에 따라 일정 기간 보관이 필요한 정보는 별도로 보관합니다.",
-            },
-        ],
-    },
-    {
-        title: "7. 이용자 및 법정대리인의 권리와 행사 방법",
-        content: [
-            `<p class="mb-4 font-bold">
+    ],
+    figures: [
+      {
+        imageSrc: ["/images/보유기간.png", "/images/파기.png"],
+        alt: "보유기간 및 파기",
+        caption: "보유기간 및 파기",
+        description:
+          "수집 및 이용 목적의 달성 또는 회원 탈퇴 등 파기사유가 발생한 개인정보는 안전하게 파기합니다. 수집된 개인정보는 사전에 안내된 보유기간 동안만 보관 후 파기하며, 관계법령에 따라 일정 기간 보관이 필요한 정보는 별도로 보관합니다.",
+      },
+    ],
+  },
+  {
+    title: "7. 이용자 및 법정대리인의 권리와 행사 방법",
+    content: [
+      `<p class="mb-4 font-bold">
          이용자는 자신의 개인정보 처리에 관하여 아래와 같은 권리를 가질 수 있습니다.
       </p>`,
-            `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
+      `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
           <li class="mb-4">개인정보 열람(조회)을 요구할 권리</li>
           <li class="mb-4">개인정보 정정을 요구할 권리</li>
           <li class="mb-4">개인정보 처리정지를 요구할 권리</li>
           <li class="mb-4">개인정보 삭제요구 및 동의철회/탈퇴를 요구할 권리</li>
        </ul>`,
-            `<div>
+      `<div>
           <p class="mt-16 mb-4 font-bold">
             이용자는 서비스 내 다음과 같은 기능을 통해 언제든지 개인정보 열람(조회) 등의 권리를 직접
             행사하거나 또는 고객센터를 통해 요청할 수 있습니다.
           </p>`,
-            `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
+      `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
           <li class="mb-4">링크드아웃계정 정보 열람 및 수정</li>
           <li class="mb-4">링크드아웃계정과 연결된 서비스 관리</li>
           <li class="mb-4">
             링크드아웃 서비스 이용을 위한 개인정보 수집, 이용, 제3자 제공 현황 확인
           </li>
        </ul>`,
-            `<p>
+      `<p>
          링크드아웃은 이용자의 요청을 받은 경우 이를 지체없이 처리하며, 이용자가 개인정보의 오류에 대한 
          정정을 요청한 경우 정정을 완료하기 전까지 해당 개인정보를 이용 또는 제공하지 않습니다.
       </p>
        </div>`,
-            `<div>
+      `<div>
           <p class="mt-16 mb-4 font-bold">
             14세 미만 아동의 개인정보를 처리할 경우에는 법정대리인의 동의를 받아야 합니다. 법정대리인은
             아동의 개인정보를 조회하거나 수정 및 삭제, 처리정지 및 동의 철회 등의 권리를 행사할 수 있습니다.
           </p>`,
-            `<p class="mb-4">
+      `<p class="mb-4">
          법정대리인 동의를 받기 위하여 아동에게 법정대리인의 성명, 연락처와 같이 최소한의 정보를 요구할 
          수 있으며, 아래 방법으로 법정대리인의 동의를 확인합니다.
       </p>`,
-            `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
+      `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
           <li class="mb-4">법정대리인의 휴대전화 본인인증을 통해 본인여부를 확인하는 방법</li>
           <li class="mb-4">
             법정대리인에게 동의내용이 적힌 서면을 제공하여 서명날인 후 제출하게 하는 방법
@@ -405,25 +403,25 @@ const policiesData = [
           </li>
        </ul>
        </div>`,
-        ],
-        figures: [
-            {
-                imageSrc: ["/images/정보주체의권리의무.png", "/images/열람청구.png"],
-                alt: "정보주체의권리의무",
-                caption: "정보주체의 권리의무",
-                description:
-                    "링크드아웃 이용자는 언제든지 자신의 개인정보를 조회하거나 수정할 수 있으며, 동의한 개인정보 처리에 대해 동의 철회를 요구할 수 있습니다.",
-            },
-        ],
-    },
-    {
-        title: "8. 개인정보 자동 수집 장치에 관한 사항",
-        content: [
-            `<p class="mb-4 font-bold">웹기반 서비스 제공을 위하여 쿠키를 설치・운영할 수 있습니다.</p>`,
-            `<p class="mb-4">
+    ],
+    figures: [
+      {
+        imageSrc: ["/images/정보주체의권리의무.png", "/images/열람청구.png"],
+        alt: "정보주체의권리의무",
+        caption: "정보주체의 권리의무",
+        description:
+          "링크드아웃 이용자는 언제든지 자신의 개인정보를 조회하거나 수정할 수 있으며, 동의한 개인정보 처리에 대해 동의 철회를 요구할 수 있습니다.",
+      },
+    ],
+  },
+  {
+    title: "8. 개인정보 자동 수집 장치에 관한 사항",
+    content: [
+      `<p class="mb-4 font-bold">웹기반 서비스 제공을 위하여 쿠키를 설치・운영할 수 있습니다.</p>`,
+      `<p class="mb-4">
          빠르고 편리한 웹사이트 사용을 지원하고 맞춤형 서비스를 제공하기 위해 쿠키를 사용합니다.
       </p>`,
-            `<div class="grid grid-cols-1 gap-4">
+      `<div class="grid grid-cols-1 gap-4">
           <div class="my-8">
               <h3 class="mb-6 font-bold">쿠키란?</h3>
               <p>
@@ -431,7 +429,7 @@ const policiesData = [
                 파일로 이용자 PC에 저장됩니다.
               </p>
           </div>`,
-            `<div class="my-4">
+      `<div class="my-4">
           <h3 class="mb-6 font-bold">사용목적</h3>
           <p>
             개인화되고 맞춤화된 서비스를 제공하기 위해서 이용자의 정보를 저장하고 수시로 불러오는 쿠키를 
@@ -441,7 +439,7 @@ const policiesData = [
             있도록 돕습니다.
           </p>
        </div>`,
-            `<div class="my-8">
+      `<div class="my-8">
           <h3 class="mb-6 font-bold">쿠키 수집 거부</h3>
           <p>
             이용자는 쿠키 설치에 대한 선택권을 가지고 있으며, 웹브라우저 상단의 &lsquo;설정 &gt; 
@@ -477,14 +475,14 @@ const policiesData = [
               </ul>
           </div>
        </div>`,
-        ],
-        figures: null,
-    },
-    {
-        title: "9. 개인정보의 안전성 확보 조치에 관한 사항",
-        content: [
-            `<p class="mb-4 font-bold">링크드아웃은 이용자의 개인정보 보호를 위해 아래의 노력을 합니다.</p>`,
-            `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
+    ],
+    figures: null,
+  },
+  {
+    title: "9. 개인정보의 안전성 확보 조치에 관한 사항",
+    content: [
+      `<p class="mb-4 font-bold">링크드아웃은 이용자의 개인정보 보호를 위해 아래의 노력을 합니다.</p>`,
+      `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
           <li class="mb-4">
             이용자의 개인정보를 암호화된 통신구간을 이용하여 전송하고, 비밀번호 등 중요정보는 암호화하여 
             보관하고 있습니다.
@@ -500,7 +498,7 @@ const policiesData = [
             감사를 실시하고 있습니다.
           </li>
        </ul>`,
-            `<div class="p-4 text-center border rounded-lg info-card">
+      `<div class="p-4 text-center border rounded-lg info-card">
           <figure class="flex flex-col items-center">
             <div class="w-[100px] h-[100px] mb-8">
                 <Image src="/images/안정성확보조치.png" alt="안정성 확보조치" width={100} height={100} />
@@ -512,21 +510,21 @@ const policiesData = [
             있으며, 보호조치 수준에 대해서는 내부 점검 및 외부 검증을 통해 정기적으로 확인하고 있습니다.
           </p>
        </div>`,
-            `<p class="my-4 font-bold">
+      `<p class="my-4 font-bold">
          링크드아웃은 유럽연합 일반 개인정보보호법(General Data Protection Regulation) 및 EU 회원국의 
          법률(이하 &lsquo;GDPR등&rsquo;)을 준수합니다.
       </p>`,
-            `<p class="mb-4">
+      `<p class="mb-4">
          유럽연합 내 이용자를 대상으로 서비스를 제공하는 경우, 아래 내용이 적용될 수 있습니다.
       </p>`,
-            `<div>
+      `<div>
           <h3 class="mt-12 mb-4 font-semibold">[개인정보 처리의 목적 및 처리근거]</h3>
           <p class="mb-4">
             링크드아웃은 수집한 개인정보를 &ldquo;3. 개인정보의 이용&rdquo;에 기재된 목적으로만 이용하며, 
             사전에 이용자에게 해당 사실을 알리고 동의를 구하고 있습니다. 그리고 GDPR등에 따라 링크드아웃은 
             아래 하나에 해당하는 경우에 이용자의 개인정보를 처리할 수 있습니다.
           </p>`,
-            `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
+      `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
           <li class="mb-4">정보주체의 동의</li>
           <li class="mb-4">정보주체와의 계약의 체결 및 이행을 위한 경우</li>
           <li class="mb-4">법적 의무사항 준수를 위한 경우</li>
@@ -537,7 +535,7 @@ const policiesData = [
           </li>
        </ul>
        </div>`,
-            `<div>
+      `<div>
           <h3 class="mt-12 mb-4 font-semibold">[유럽연합 내 이용자의 권리 보장]</h3>
           <p class="mb-4">
             GDPR 등에 따라 이용자는 자신의 개인정보를 다른 관리자에게 이전해 달라고 요청할 수 있고, 자신의 
@@ -546,27 +544,27 @@ const policiesData = [
             수 있으며, 사전에 이용자의 동의를 구하고 있습니다. 이용자는 원하지 않는 경우 언제든지 동의를 
             철회할 수 있습니다.
           </p>`,
-            `<p class="mb-4">
+      `<p class="mb-4">
          위와 관련한 요청사항은 고객센터를 통해 서면, 전화 또는 이메일로 연락하시면 지체 없이 
          조치하겠습니다.
       </p>`,
-            `<p>
+      `<p class="mb-12">
          개인정보의 오류에 대한 정정을 요청하신 경우 정정을 완료하기 전까지 해당 개인정보를 이용 또는 
          제공하지 않습니다.
       </p>`,
-        ],
-        figure: [],
-    },
-    {
-        title: "10. 개인정보 보호책임자 및 고충처리 부서",
-        content: [
-            `<p class="mb-4 font-bold">
+    ],
+    figure: [],
+  },
+  {
+    title: "10. 개인정보 보호책임자 및 고충처리 부서",
+    content: [
+      `<p class="mb-4 font-bold">
          링크드아웃은 이용자의 개인정보 관련 문의사항 및 불만 처리 등을 위하여 개인정보 보호책임자를 지정하고 
          있습니다.
       </p>`,
-            `<div>
+      `<div>
           <p class="mb-4 font-bold">개인정보 보호 책임자</p>`,
-            `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
+      `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
           <li class="mb-4">책임자: 조대찬 (대표)</li>
           <li class="mb-4">문의</li>
           <li class="list-none">
@@ -578,56 +576,59 @@ const policiesData = [
               </ul>
           </li>
        </ul>`,
-            `<p class="mb-4">
+      `<p class="mb-4">
          개인정보보호 관련 문의, 불만 등 의견이 있는 경우 위 연락처로 접수해 주시면 신속하게 검토하여 
          답변 드리겠습니다.
       </p>
        </div>`,
+    ],
+    figures: [
+      {
+        imageSrc: [
+          "/images/개인정보보호책임자.png",
+          "/images/고충처리부서.png",
         ],
-        figures: [
-            {
-                imageSrc: ["/images/개인정보보호책임자.png", "/images/고충처리부서.png"],
-                alt: "개인정보보호책임자 & 고충처리 부서",
-                caption: "개인정보보호책임자 & 고충처리 부서",
-                description:
-                    "링크드아웃 서비스 이용 시 발생하는 모든 개인정보보호 관련 문의 및 불만사항은 개인정보보호책임자 고충처리 부서(링크드아웃 고객센터)로 문의해주시면 됩니다.",
-            },
-            {
-                imageSrc: "/images/권익침해구제.png",
-                alt: "권익침해 구제",
-                caption: "권익침해 구제",
-                description:
-                    "개인정보가 침해되어서 이에 대한 신고나 상담이 필요하신 경우 국가기관에 문의하여 도움을 받으실 수 있습니다.",
-            },
-        ],
-    },
-    {
-        title: "11. 개정 전 고지의무 등 안내",
-        content: [
-            `<p class="mb-4">
+        alt: "개인정보보호책임자 & 고충처리 부서",
+        caption: "개인정보보호책임자 & 고충처리 부서",
+        description:
+          "링크드아웃 서비스 이용 시 발생하는 모든 개인정보보호 관련 문의 및 불만사항은 개인정보보호책임자 고충처리 부서(링크드아웃 고객센터)로 문의해주시면 됩니다.",
+      },
+      {
+        imageSrc: "/images/권익침해구제.png",
+        alt: "권익침해 구제",
+        caption: "권익침해 구제",
+        description:
+          "개인정보가 침해되어서 이에 대한 신고나 상담이 필요하신 경우 국가기관에 문의하여 도움을 받으실 수 있습니다.",
+      },
+    ],
+  },
+  {
+    title: "11. 개정 전 고지의무 등 안내",
+    content: [
+      `<p class="mb-4">
          법령이나 서비스의 변경사항을 반영하기 위한 목적 등으로 개인정보 처리방침을 수정할 수 있습니다. 
          개인정보 처리방침이 변경되는 경우 최소 7일 전 변경 사항을 사전에 안내 하겠습니다. 다만, 이용자 
          권리의 중대한 변경이 발생할 때에는 최소 30일 전에 미리 알려드리겠습니다.
       </p>`,
-            `<p class="mb-4">
+      `<p class="mb-4">
          링크드아웃은 이용자 여러분의 정보를 소중히 생각하며, 이용자가 더욱 안심하고 서비스를 이용할 수 
          있도록 최선의 노력을 다할 것을 약속드립니다.
       </p>`,
-            `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
+      `<ul class="pl-4 mb-4 ml-4 list-disc sm:pl-6">
           <li class="mb-4">공고일자: 2024년 00월 00일</li>
           <li>시행일자: 2024년 00월 00일</li>
        </ul>`,
-        ],
-        figures: [
-            {
-                imageSrc: "/images/개인정보처리방침변경.png",
-                alt: "개인정보처리방침변경",
-                caption: "개인정보처리방침변경",
-                description:
-                    "링크드아웃 서비스 이용 시 발생하는 모든 개인정보보호 관련 문의 및 불만사항은 개인정보보호책임자 고충처리 부서(링크드아웃 고객센터)로 문의해주시면 됩니다.",
-            },
-        ],
-    },
-]
+    ],
+    figures: [
+      {
+        imageSrc: "/images/개인정보처리방침변경.png",
+        alt: "개인정보처리방침변경",
+        caption: "개인정보처리방침변경",
+        description:
+          "링크드아웃 서비스 이용 시 발생하는 모든 개인정보보호 관련 문의 및 불만사항은 개인정보보호책임자 고충처리 부서(링크드아웃 고객센터)로 문의해주시면 됩니다.",
+      },
+    ],
+  },
+];
 
-export default policiesData
+export default policiesData;
