+++ homepage/app/globals.css
@@ -2,54 +2,95 @@
 @tailwind components;
 @tailwind utilities;
 
-/* Responsive styling */
+/* 스타일을 적용할 특정 섹션의 클래스 */
+.first-section-grid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 1rem;
+}
+
 @media (min-width: 640px) {
-    .policies-container {
-        padding: 2rem;
-    }
-
-    .section {
-        padding: 2.5rem;
-    }
-
-    .title {
-        font-size: 1.5rem;
-    }
-
-    .figure-grid {
-        display: flex;
-        flex-wrap: wrap;
-        gap: 20px;
-        align-items: center;
-        margin: 10px;
-        border: 1px solid #ccc;
-        border-radius: 5px;
-    }
-
-    .figure-item {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        text-align: center;
-        justify-content: center;
-    }
+  .first-section-grid {
+    grid-template-columns: repeat(3, 1fr);
+  }
 }
 
-@media (min-width: 768px) {
-    .figure-grid {
-        flex-direction: row;
-    }
+@media (min-width: 1024px) {
+  .first-section-grid {
+    grid-template-columns: repeat(6, 1fr);
+  }
 }
 
-/* 정책 페이지 border */
-.border-right:not(:last-child) {
-    border-right: 1px solid #ccc;
-}
-
-.border-right:last-child {
-    border-right: none;
-}
+/* 섹션 간 여백을 추가 */
 
 .section:last-child {
-    border-bottom: none;
+  margin-bottom: 0;
+}
+
+.policies-container {
+  padding-left: 1rem;
+  padding-right: 1rem;
+}
+
+@media (min-width: 640px) {
+  .policies-container {
+    padding: 2rem;
+  }
+
+  .section {
+    padding: 2.5rem;
+  }
+
+  .title {
+    font-size: 1.5rem;
+  }
+}
+
+@media (max-width: 640px) {
+  .policies-container {
+    padding-left: 1rem;
+    padding-right: 1rem;
+  }
+
+  .section {
+    padding-left: 1.5rem;
+    padding-right: 1.5rem;
+    margin-bottom: 3rem;
+  }
+
+  .title {
+    font-size: 1.25rem;
+  }
+
+  .figure-grid {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    align-items: stretch;
+    margin: 10px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+  }
+
+  .figure-item {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+    justify-content: center;
+    width: 100%;
+  }
+}
+
+/* figure 중앙 정렬 */
+.figure-item {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  justify-content: center;
+}
+
+.figure-item img {
+  margin-bottom: 0.5rem;
 }
