+++ homepage/app/policy/page.jsx
@@ -1,19 +1,25 @@
-import { Section } from "../_components"
-import policiesData from "../data/policiesData"
+import { Policy } from "../_components";
+import policiesData from "../data/policiesData";
 
 const Index = () => {
-    return (
-        <>
-            <div>
-                {policiesData.map((policy, index) => (
-                    <Section key={index} title={policy.title} content={policy.content} figures={policy.figures} />
-                ))}
-            </div>
-            <footer className="mx-6 my-8 text-sm text-grey-500">
-                본 서비스의 모든 저작권은 링크드아웃에 있으며, 어떠한 형태로든 복제나 배포를 금지합니다.
-            </footer>
-        </>
-    )
-}
+  return (
+    <>
+      <div className="policies-container">
+        {policiesData.map((policy, index) => (
+          <Policy
+            key={index}
+            title={policy.title}
+            content={policy.content}
+            figures={policy.figures}
+          />
+        ))}
+      </div>
+      <footer className="mx-6 my-8 text-sm text-grey-500">
+        본 서비스의 모든 저작권은 링크드아웃에 있으며, 어떠한 형태로든 복제나
+        배포를 금지합니다.
+      </footer>
+    </>
+  );
+};
 
-export default Index
+export default Index;
