+++ homepage/package.json
@@ -9,14 +9,15 @@
         "lint": "next lint"
     },
     "dependencies": {
+        "autoprefixer": "^10.4.20",
+        "next": "^13.4.0",
         "react": "^18",
-        "react-dom": "^18",
-        "next": "^13.4.0"
+        "react-dom": "^18"
     },
     "devDependencies": {
-        "postcss": "^8",
-        "tailwindcss": "^3.4.1",
         "eslint": "^8",
-        "eslint-config-next": "14.2.5"
+        "eslint-config-next": "14.2.5",
+        "postcss": "^8",
+        "tailwindcss": "^3.4.1"
     }
 }
